adding improves in search directory and screenshot pages

+++ src/Pages/Screenshot.js
@@ -3,14 +3,19 @@ import axios from 'axios';
 import React from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import { Button } from '@mui/material';
-
-
+import ImageSelector from '../Components/ImageSelector';
 
 const ButtonStyles = {
     "margin": "20px 5px"
 }
 
 const Screenshot = (props) => {
+
+
+    const handleSetImage =  () => {
+        console.log("clicked use image");
+
+    }
 
     const api_url = process.env.REACT_APP_AXIOS_BASE_URL;
 
@@ -33,34 +38,42 @@ const Screenshot = (props) => {
 
 
     const deleteScreenshot = () => {
-        axios.delete(api_url + "/api/screenshots/" + screenshotId.id).then(function (response) {
-            console.log(response.data);
-            navigate('/');
-        })
-            .catch(function (error) {
-                console.log(error);
+
+        let result = window.confirm("Are you sure to delete the Screenshot?");
+
+
+        if (result) {
+            axios.delete(api_url + "/api/screenshots/" + screenshotId.id).then(function (response) {
+                console.log(response.data);
+                navigate('/');
             })
-            .then(function () {
-                // always executed
-            });
+                .catch(function (error) {
+                    console.log(error);
+                })
+
+        }
+
+        else {
+            console.log('alert screentho not deleted');
+        }
+
     }
 
     const image_src = api_url + "/images/" + apiResponse.hostname + "/" + apiResponse.title + ".jpg"
     return (
         <div className="Maincontainer">
-
-
             <div className='ScreenshotContainer'>
                 <img src={image_src} />
-
                 <div className='ActionsContainer'>
+                    <Button onClick={()=> handleSetImage} style={ButtonStyles} variant="contained">Use Image</Button>
 
-                    <Button style={ButtonStyles} variant="contained">Use Image</Button>
+
                     <Button style={ButtonStyles} onClick={deleteScreenshot} variant="contained" color="error">Delete</Button>
                 </div>
 
-            </div>
 
+                <ImageSelector></ImageSelector>
+            </div>
         </div>
     );
 }
